+++ docs/src/pages/_meta.tsx
@@ -1,0 +1,10 @@
+const meta = {
+  index: 'Projects',
+  assert: '@httpx/assert',
+  exception: '@httpx/exception',
+  'dsn-parser': '@httpx/dsn-parser',
+  'plain-object': '@httpx/plain-object',
+  sponsors: 'Sponsors',
+  license: 'License',
+};
+export default meta;