+++ integrations/prisma-exception/.eslintrc.cjs
@@ -17,7 +17,7 @@ module.exports = {
     '@belgattitude/eslint-config-bases/import-x',
     '@belgattitude/eslint-config-bases/sonar',
     '@belgattitude/eslint-config-bases/regexp',
-    '@belgattitude/eslint-config-bases/jest',
+    '@belgattitude/eslint-config-bases/vitest',
     '@belgattitude/eslint-config-bases/performance',
     // Apply prettier and disable incompatible rules
     '@belgattitude/eslint-config-bases/prettier-plugin',
@@ -35,7 +35,7 @@ module.exports = {
     {
       files: ['src/**/*.ts'],
       rules: {
-        'import-x/no-cycle': [1, { maxDepth: 100 }],
+        // 'import-x/no-cycle': [1, { maxDepth: 100 }],
         'import-x/no-extraneous-dependencies': [
           'error',
           {