+++ packages/dsn-parser/src/is-parsable-dsn.ts
@@ -2,6 +2,5 @@ import type { ParsableDsn } from './dsn-parser.type';
 import { parseDsn } from './parse-dsn';
 
 export const isParsableDsn = (dsn: unknown): dsn is ParsableDsn => {
-  const parsed = parseDsn(dsn as string);
-  return parsed.success;
+  return parseDsn(dsn as string).success;
 };