+++ packages/memo-intl/vitest.config.ts
@@ -12,8 +12,14 @@ export default defineConfig({
     target: ['node18'],
   },
   plugins: [tsconfigPaths(), ...[cspeed].filter(Boolean)],
-  cacheDir: '../../.cache/vite/httpx-dsn-parser',
+  cacheDir: '../../.cache/vite/memo-intl',
   test: {
+    browser: {
+      provider: 'playwright', // or 'webdriverio'
+      enabled: false,
+      // at least one instance is required
+      instances: [{ browser: 'chromium' }],
+    },
     // @link https://vitest.dev/config/#clearmocks
     clearMocks: true,
     coverage: {