+++ packages/plain-object/bench/comparative.bench.ts
@@ -1,29 +1,30 @@
 import { bench, describe } from 'vitest';
 
+import { devDependencies } from '../package.json' with { type: 'json' };
+
+const versions = devDependencies;
+
 /**
  * Based on a hypothesis:
- * `isPlainObject` is called 70% of the time with valid plain objects,
- * 10% with maps, 10% with nulls, 5% with undefined and 5% with string.
  */
 const realLifeScenarios = [
-  // 70% plain object
-  ...Array.from({ length: 35 }).map((_) => ({
-    key1: Math.random(),
-    key2: Math.random(),
-    key3: Math.random(),
-  })),
-  ...Array.from({ length: 35 }).map((_) => ({
+  ...Array.from({ length: 10 }).map((_) => ({})),
+  ...Array.from({ length: 10 }).map((_) => ({
     key1: {
       subkey: [],
     },
   })),
   // Others: not plain objects
-  ...Array.from({ length: 10 }).fill(new Map([['key', Math.random()]])),
-  ...Array.from({ length: 5 }).fill(null),
-  ...Array.from({ length: 5 }).fill([]),
+  ...Array.from({ length: 10 }).fill(new Map([['key', 10]])),
+  ...Array.from({ length: 10 }).fill(new Date()),
+  ...Array.from({ length: 10 }).fill(null),
   // eslint-disable-next-line unicorn/no-useless-undefined
-  ...Array.from({ length: 5 }).fill(undefined),
-  ...Array.from({ length: 5 }).fill('str'),
+  ...Array.from({ length: 10 }).fill(undefined),
+  ...Array.from({ length: 10 }).fill(1),
+  ...Array.from({ length: 10 }).fill(0),
+  ...Array.from({ length: 10 }).fill('str'),
+  ...Array.from({ length: 10 }).fill(''),
+  ...Array.from({ length: 10 }).fill(Number.NaN),
 ] as const;
 
 describe(`Compare calling isPlainObject with ${realLifeScenarios.length}x mixed types values`, async () => {
@@ -53,6 +54,10 @@ describe(`Compare calling isPlainObject with ${realLifeScenarios.length}x mixed 
     (mod) => mod.isPlainObject
   );
 
+  const reduxIsPlainObject = await import('redux').then(
+    (mod) => mod.isPlainObject
+  );
+
   // @ts-expect-error packaging of this lib is not compatible with latest ts / module
   // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
   const jonschlinkertIsPlainObject = await import('is-plain-object').then(
@@ -60,39 +65,50 @@ describe(`Compare calling isPlainObject with ${realLifeScenarios.length}x mixed 
     (mod) => mod.isPlainObject
   );
 
-  bench('@httpx/plain-object: `isPlainObject(v)`', () => {
+  bench('"@httpx/plain-object": `isPlainObject(v)`', () => {
     for (const value of realLifeScenarios) {
       httpxIsPlainObject(value);
     }
   });
 
-  bench('(sindresorhus/)is-plain-obj: `isPlainObj(v)`', () => {
+  bench(`"is-plain-obj":"${versions['is-plain-obj']}": 'isPlainObj(v)'`, () => {
     for (const value of realLifeScenarios) {
       isPlainObj(value);
     }
   });
 
-  bench('@sindresorhus/is: `is.plainObject(v)`', () => {
+  bench(
+    `"@sindresorhus/is":"${versions['@sindresorhus/is']}": 'is.plainObject(v)'`,
+    () => {
+      for (const value of realLifeScenarios) {
+        is.plainObject(value);
+      }
+    }
+  );
+
+  bench(`"es-toolkit":"${versions['es-toolkit']}": 'isPlainObject(v)'`, () => {
     for (const value of realLifeScenarios) {
-      is.plainObject(value);
+      esToolkitIsPlainObject(value);
     }
   });
 
-  bench('estoolkit:  `isPlainObject(v)`', () => {
+  bench(`"redux":"${versions.redux}": 'isPlainObject(v)'`, () => {
     for (const value of realLifeScenarios) {
-      // Sadly es-toolkit does not give a very usable type for isPlainObject
-      esToolkitIsPlainObject(value as object);
+      reduxIsPlainObject(value);
     }
   });
 
-  bench('(jonschlinkert/)is-plain-object: `isPlainObject(v)`', () => {
-    for (const value of realLifeScenarios) {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-      jonschlinkertIsPlainObject(value);
+  bench(
+    `"is-plain-object":"${versions['is-plain-object']}": 'isPlainObject(v)'`,
+    () => {
+      for (const value of realLifeScenarios) {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-call
+        jonschlinkertIsPlainObject(value);
+      }
     }
-  });
+  );
 
-  bench('lodash-es: `_.isPlainObject(v)`', () => {
+  bench(`lodash-es:"${versions['lodash-es']}": '_.isPlainObject(v)'`, () => {
     for (const value of realLifeScenarios) {
       lodashIsPlainObject(value);
     }