+++ packages/plain-object/src/index.ts
@@ -1,3 +1,5 @@
 export { assertPlainObject } from './assert-plain-object';
 export { isPlainObject } from './is-plain-object';
+export type { StaticBuiltInClass } from './is-static-built-in-class';
+export { isStaticBuiltInClass } from './is-static-built-in-class';
 export type { PlainObject } from './plain-object.types';