+++ packages/plain-object/src/is-plain-object.ts
@@ -4,7 +4,7 @@ import type { PlainObject } from './plain-object.types';
 /**
  * Check if a value is a plain object
  *
- * An object is plain if it's created by either {}, new Object(), or Object.create(null).
+ * A plain object is a basic JavaScript object, such as {}, { data: [] }, new Object() or Object.create(null).
  *
  * @example
  * ```typescript
@@ -12,6 +12,7 @@ import type { PlainObject } from './plain-object.types';
  *
  * // ✅👇 True
  *
+ * isPlainObject({ });                       // ✅
  * isPlainObject({ key: 'value' });          // ✅
  * isPlainObject({ key: new Date() });       // ✅
  * isPlainObject(new Object());              // ✅
@@ -35,10 +36,20 @@ import type { PlainObject } from './plain-object.types';
  * isPlainObject('hello');             // ❌
  * isPlainObject([]);                  // ❌
  * isPlainObject(new Date());          // ❌
- * isPlainObject(Math);                // ❌ Static built-in classes
+ * isPlainObject(new Uint8Array([1])); // ❌
+ * isPlainObject(Buffer.from('ABC'));  // ❌
  * isPlainObject(Promise.resolve({})); // ❌
  * isPlainObject(Object.create({}));   // ❌
+ * isPlainObject(new (class Cls {}));  // ❌
+ * isPlainObject(globalThis);          // ❌,
  * ```
+ *
+ * // ✅👇 Note that static built-in classes are treated as plain objects
+ * //    check for `isStaticBuiltInClass` to exclude if needed
+ *
+ * isPlainObject(Math);                // ✅
+ * isPlainObject(JSON);                // ✅
+ * isPlainObject(Atomics);             // ✅
  */
 export const isPlainObject = <
   TValue extends BasePlainObject = DefaultBasePlainObject,
@@ -53,12 +64,9 @@ export const isPlainObject = <
 
   const proto = Object.getPrototypeOf(v) as typeof Object.prototype | null;
   return (
-    (proto === null ||
-      proto === Object.prototype ||
-      // Required to support node:vm.runInNewContext({})
-      Object.getPrototypeOf(proto) === null) &&
-    // https://stackoverflow.com/a/76387885/5490184
-    !(Symbol.toStringTag in v) &&
-    !(Symbol.iterator in v)
+    proto === null ||
+    proto === Object.prototype ||
+    // Required to support node:vm.runInNewContext({})
+    Object.getPrototypeOf(proto) === null
   );
 };