+++ packages/treeu/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @httpx/treeu
 
+## 0.4.8
+
+### Patch Changes
+
+- Updated dependencies [[`946f440`](https://github.com/belgattitude/httpx/commit/946f4400576ccc72f1b3c9fa140e84ea82e0e997)]:
+  - @httpx/plain-object@2.0.4
+
+## 0.4.7
+
+### Patch Changes
+
+- Updated dependencies [[`be1dbb3`](https://github.com/belgattitude/httpx/commit/be1dbb30bca671124f91de5714aadfc0a46789a0)]:
+  - @httpx/plain-object@2.0.3
+
+## 0.4.6
+
+### Patch Changes
+
+- [#1546](https://github.com/belgattitude/httpx/pull/1546) [`bdf9e19`](https://github.com/belgattitude/httpx/commit/bdf9e19d11bc66b9b7279da8f292a889f0acbffa) Thanks [@belgattitude](https://github.com/belgattitude)! - Ensure CI tests on Clouflare workers and latest chrome (playwright)
+
+- Updated dependencies [[`bdf9e19`](https://github.com/belgattitude/httpx/commit/bdf9e19d11bc66b9b7279da8f292a889f0acbffa)]:
+  - @httpx/plain-object@2.0.2
+
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies [[`d1a01e2`](https://github.com/belgattitude/httpx/commit/d1a01e2af290f9f5bc0f2357773616b39a9e048d)]:
+  - @httpx/plain-object@2.0.1
+
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies [[`4e1ecf4`](https://github.com/belgattitude/httpx/commit/4e1ecf4088664896e0d43ee765efa6f6b270054b), [`4e1ecf4`](https://github.com/belgattitude/httpx/commit/4e1ecf4088664896e0d43ee765efa6f6b270054b), [`4e1ecf4`](https://github.com/belgattitude/httpx/commit/4e1ecf4088664896e0d43ee765efa6f6b270054b), [`4e1ecf4`](https://github.com/belgattitude/httpx/commit/4e1ecf4088664896e0d43ee765efa6f6b270054b), [`4e1ecf4`](https://github.com/belgattitude/httpx/commit/4e1ecf4088664896e0d43ee765efa6f6b270054b), [`4e1ecf4`](https://github.com/belgattitude/httpx/commit/4e1ecf4088664896e0d43ee765efa6f6b270054b), [`4e1ecf4`](https://github.com/belgattitude/httpx/commit/4e1ecf4088664896e0d43ee765efa6f6b270054b)]:
+  - @httpx/plain-object@2.0.0
+
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies [[`003cdb3`](https://github.com/belgattitude/httpx/commit/003cdb3f4e0ff00970ac00b56659a47a30f7358f)]:
+  - @httpx/plain-object@1.2.0
+
 ## 0.4.2
 
 ### Patch Changes